+++ Package.swift
@@ -4,6 +4,10 @@ import PackageDescription
 
 let package = Package(
     name: "Ether",
+    products: [
+        .executable(name: "Executable", targets: ["Executable"]),
+        .library(name: "Ether", targets: ["Helpers", "Ether"])
+    ],
     dependencies: [
         .package(url: "https://github.com/vapor/vapor.git", from: "3.0.3"),
         .package(url: "https://github.com/vapor/console.git", from: "3.0.2"),